Make courses user specific
Courses will only show up for the user that added them. This change added
a foreign key field to the course model in order to keep track of the
course's owner. Anonymous users cannot add classes and the UI on the index
page hides the 'Add Class' button for them.