Fix new setups
A brand new clone of the repository would not launch. First, there was
reference to a local file in requirements.txt. Second, a missing comma
made django import 'whitenoisedjango' instead of 'whitenoise' which caused
an import error. Third, the initalization code for CourseForm refered to
the database. Now courses are created without any option to add fields at
first.